Skip delivery if folder does not exist (#228)
This pull request makes a minor improvement to the `DeliverAppFile.ps1`
script by adding a check to ensure each target folder exists before
attempting to push it to the NuGet feed. This helps prevent errors if a
specified folder is missing.

- Reliability improvement:
* Added a check using `Test-Path` to confirm that each `$folderName`
exists before calling `Push-AppToNuGetFeed`, reducing the chance of
errors during the delivery process.