Fix read settings for setup pipeline step (#239)
This pull request introduces a minor enhancement to the settings reading
logic in `ReadSettings/ReadSettings.ps1`. The main change is the
addition of a conditional check to identify artifacts only when the
pipeline is not `SetupPipelines`.

Settings handling improvements:

* Added a condition to skip artifact identification logic if the
environment variable `AL_PIPELINENAME` is set to `SetupPipelines`,
ensuring artifact detection only occurs for relevant pipelines.
* Moved the closing brace for the `try` block to properly encapsulate
the new conditional logic.